+++ lib/src/public/SuperwallPlacementInfo.dart
@@ -4,20 +4,20 @@ import 'package:superwallkit_flutter/superwallkit_flutter.dart';
 
 /// Contains information about the internally tracked superwall placement.
 class SuperwallEventInfo {
-  SuperwallPlacement event;
+  SuperwallEvent event;
   Map<dynamic, dynamic>? params;
 
   SuperwallEventInfo({required this.event, this.params});
 
   factory SuperwallEventInfo.fromJson(Map<dynamic, dynamic> json) {
     return SuperwallEventInfo(
-      event: SuperwallPlacement.fromJson(json['placement']),
+      event: SuperwallEvent.fromJson(json['placement']),
       params: json['params'],
     );
   }
 }
 
-enum PlacementType {
+enum EventType {
   firstSeen,
   appOpen,
   appLaunch,
@@ -75,8 +75,8 @@ enum PlacementType {
   shimmerViewComplete
 }
 
-class SuperwallPlacement {
-  final PlacementType type;
+class SuperwallEvent {
+  final EventType type;
 
   // Optional per placement type
   final String? placementName;
@@ -100,7 +100,7 @@ class SuperwallPlacement {
   final Map<dynamic, dynamic>? userAttributes;
   final String? token;
 
-  SuperwallPlacement._(
+  SuperwallEvent._(
       {required this.type,
       this.placementName,
       this.deviceAttributes,
@@ -123,185 +123,182 @@ class SuperwallPlacement {
       this.userAttributes,
       this.token});
 
-  factory SuperwallPlacement.fromJson(Map<dynamic, dynamic> json) {
+  factory SuperwallEvent.fromJson(Map<dynamic, dynamic> json) {
     switch (json['placement']) {
       case 'firstSeen':
-        return SuperwallPlacement._(type: PlacementType.firstSeen);
+        return SuperwallEvent._(type: EventType.firstSeen);
       case 'appOpen':
-        return SuperwallPlacement._(type: PlacementType.appOpen);
+        return SuperwallEvent._(type: EventType.appOpen);
       case 'appLaunch':
-        return SuperwallPlacement._(type: PlacementType.appLaunch);
+        return SuperwallEvent._(type: EventType.appLaunch);
       case 'identityAlias':
-        return SuperwallPlacement._(type: PlacementType.identityAlias);
+        return SuperwallEvent._(type: EventType.identityAlias);
       case 'appInstall':
-        return SuperwallPlacement._(type: PlacementType.appInstall);
+        return SuperwallEvent._(type: EventType.appInstall);
       case 'sessionStart':
-        return SuperwallPlacement._(type: PlacementType.sessionStart);
+        return SuperwallEvent._(type: EventType.sessionStart);
       case 'deviceAttributes':
-        return SuperwallPlacement._(
-            type: PlacementType.deviceAttributes,
+        return SuperwallEvent._(
+            type: EventType.deviceAttributes,
             deviceAttributes: json['attributes']);
       case 'subscriptionStatusDidChange':
-        return SuperwallPlacement._(
-            type: PlacementType.subscriptionStatusDidChange);
+        return SuperwallEvent._(type: EventType.subscriptionStatusDidChange);
       case 'appClose':
-        return SuperwallPlacement._(type: PlacementType.appClose);
+        return SuperwallEvent._(type: EventType.appClose);
       case 'deepLink':
         var url = json['url'] as String?;
-        return SuperwallPlacement._(
-            type: PlacementType.deepLink, deepLinkUrl: url);
+        return SuperwallEvent._(type: EventType.deepLink, deepLinkUrl: url);
       case 'triggerFire':
-        return SuperwallPlacement._(
-          type: PlacementType.triggerFire,
+        return SuperwallEvent._(
+          type: EventType.triggerFire,
           placementName: json['placementName'] as String?,
           result: TriggerResult.fromJson(json['result']),
         );
       case 'paywallOpen':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallOpen,
+        return SuperwallEvent._(
+          type: EventType.paywallOpen,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallClose':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallClose,
+        return SuperwallEvent._(
+          type: EventType.paywallClose,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallDecline':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallDecline,
+        return SuperwallEvent._(
+          type: EventType.paywallDecline,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'transactionStart':
-        return SuperwallPlacement._(
-          type: PlacementType.transactionStart,
+        return SuperwallEvent._(
+          type: EventType.transactionStart,
           product: StoreProduct.fromJson(json['product']),
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'transactionFail':
-        return SuperwallPlacement._(
-          type: PlacementType.transactionFail,
+        return SuperwallEvent._(
+          type: EventType.transactionFail,
           error: json['error'],
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'transactionAbandon':
-        return SuperwallPlacement._(
-          type: PlacementType.transactionAbandon,
+        return SuperwallEvent._(
+          type: EventType.transactionAbandon,
           product: StoreProduct.fromJson(json['product']),
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'transactionComplete':
         final transactionJson = json['transaction'];
 
-        return SuperwallPlacement._(
-            type: PlacementType.transactionComplete,
+        return SuperwallEvent._(
+            type: EventType.transactionComplete,
             transaction: (transactionJson != null)
                 ? StoreTransaction.fromJson(transactionJson)
                 : null,
             product: StoreProduct.fromJson(json['product']),
             paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']));
       case 'subscriptionStart':
-        return SuperwallPlacement._(
-          type: PlacementType.subscriptionStart,
+        return SuperwallEvent._(
+          type: EventType.subscriptionStart,
           product: StoreProduct.fromJson(json['product']),
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'freeTrialStart':
-        return SuperwallPlacement._(
-          type: PlacementType.freeTrialStart,
+        return SuperwallEvent._(
+          type: EventType.freeTrialStart,
           product: StoreProduct.fromJson(json['product']),
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'transactionRestore':
-        return SuperwallPlacement._(
-          type: PlacementType.transactionRestore,
+        return SuperwallEvent._(
+          type: EventType.transactionRestore,
           restoreType: RestoreType.fromJson(json['restoreType']),
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'transactionTimeout':
-        return SuperwallPlacement._(
-          type: PlacementType.transactionTimeout,
+        return SuperwallEvent._(
+          type: EventType.transactionTimeout,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'userAttributes':
-        return SuperwallPlacement._(
-            type: PlacementType.userAttributes,
-            userAttributes: json['attributes']);
+        return SuperwallEvent._(
+            type: EventType.userAttributes, userAttributes: json['attributes']);
       case 'nonRecurringProductPurchase':
-        return SuperwallPlacement._(
-          type: PlacementType.nonRecurringProductPurchase,
+        return SuperwallEvent._(
+          type: EventType.nonRecurringProductPurchase,
           product: StoreProduct.fromJson(json['product']),
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallResponseLoadStart':
-        return SuperwallPlacement._(
-            type: PlacementType.paywallResponseLoadStart,
+        return SuperwallEvent._(
+            type: EventType.paywallResponseLoadStart,
             triggeredPlacementName: json['triggeredPlacementName']);
       case 'paywallResponseLoadNotFound':
-        return SuperwallPlacement._(
-            type: PlacementType.paywallResponseLoadNotFound,
+        return SuperwallEvent._(
+            type: EventType.paywallResponseLoadNotFound,
             triggeredPlacementName: json['triggeredPlacementName']);
       case 'paywallResponseLoadFail':
-        return SuperwallPlacement._(
-            type: PlacementType.paywallResponseLoadFail,
+        return SuperwallEvent._(
+            type: EventType.paywallResponseLoadFail,
             triggeredPlacementName: json['triggeredPlacementName']);
       case 'paywallResponseLoadComplete':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallResponseLoadComplete,
+        return SuperwallEvent._(
+          type: EventType.paywallResponseLoadComplete,
           triggeredPlacementName: json['triggeredPlacementName'],
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallWebviewLoadStart':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallWebviewLoadStart,
+        return SuperwallEvent._(
+          type: EventType.paywallWebviewLoadStart,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallWebviewLoadFail':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallWebviewLoadFail,
+        return SuperwallEvent._(
+          type: EventType.paywallWebviewLoadFail,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallWebviewLoadComplete':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallWebviewLoadComplete,
+        return SuperwallEvent._(
+          type: EventType.paywallWebviewLoadComplete,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallWebviewLoadTimeout':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallWebviewLoadTimeout,
+        return SuperwallEvent._(
+          type: EventType.paywallWebviewLoadTimeout,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallWebviewLoadFallback':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallWebviewLoadFallback,
+        return SuperwallEvent._(
+          type: EventType.paywallWebviewLoadFallback,
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallProductsLoadStart':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallProductsLoadStart,
+        return SuperwallEvent._(
+          type: EventType.paywallProductsLoadStart,
           triggeredPlacementName: json['triggeredPlacementName'],
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallProductsLoadFail':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallProductsLoadFail,
+        return SuperwallEvent._(
+          type: EventType.paywallProductsLoadFail,
           triggeredPlacementName: json['triggeredPlacementName'],
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'paywallProductsLoadComplete':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallProductsLoadComplete,
+        return SuperwallEvent._(
+          type: EventType.paywallProductsLoadComplete,
           triggeredPlacementName: json['triggeredPlacementName'],
         );
       case 'paywallProductsLoadRetry':
-        return SuperwallPlacement._(
-          type: PlacementType.paywallProductsLoadRetry,
+        return SuperwallEvent._(
+          type: EventType.paywallProductsLoadRetry,
           triggeredPlacementName: json['triggeredPlacementName'],
           attempt: json['attempt'],
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'surveyResponse':
-        return SuperwallPlacement._(
-          type: PlacementType.surveyResponse,
+        return SuperwallEvent._(
+          type: EventType.surveyResponse,
           survey: Survey.fromJson(json['survey']),
           selectedOption: SurveyOption.fromJson(json['selectedOption']),
           customResponse: json['customResponse'],
@@ -310,49 +307,47 @@ class SuperwallPlacement {
       case 'paywallPresentationRequest':
         final statusReasonJson = json['reason'];
 
-        return SuperwallPlacement._(
-            type: PlacementType.paywallPresentationRequest,
+        return SuperwallEvent._(
+            type: EventType.paywallPresentationRequest,
             status: PaywallPresentationRequestStatus.fromJson(json['status']),
             reason: (statusReasonJson != null)
                 ? PaywallPresentationRequestStatusReason.fromJson(
                     statusReasonJson)
                 : null);
       case 'touchesBegan':
-        return SuperwallPlacement._(type: PlacementType.touchesBegan);
+        return SuperwallEvent._(type: EventType.touchesBegan);
       case 'surveyClose':
-        return SuperwallPlacement._(type: PlacementType.surveyClose);
+        return SuperwallEvent._(type: EventType.surveyClose);
       case 'configRefresh':
-        return SuperwallPlacement._(type: PlacementType.configRefresh);
+        return SuperwallEvent._(type: EventType.configRefresh);
       case 'reset':
-        return SuperwallPlacement._(type: PlacementType.reset);
+        return SuperwallEvent._(type: EventType.reset);
       case 'customPlacement':
-        return SuperwallPlacement._(
-          type: PlacementType.customPlacement,
+        return SuperwallEvent._(
+          type: EventType.customPlacement,
           name: json['name'],
           params: json['params'],
           paywallInfo: PaywallInfo(bridgeId: json['paywallInfoBridgeId']),
         );
       case 'configAttributes':
-        return SuperwallPlacement._(type: PlacementType.configAttributes);
+        return SuperwallEvent._(type: EventType.configAttributes);
       case 'confirmAllAssignments':
-        return SuperwallPlacement._(type: PlacementType.confirmAllAssignments);
+        return SuperwallEvent._(type: EventType.confirmAllAssignments);
       case 'configFail':
-        return SuperwallPlacement._(type: PlacementType.configFail);
+        return SuperwallEvent._(type: EventType.configFail);
       case 'adServicesTokenRequestStart':
-        return SuperwallPlacement._(
-            type: PlacementType.adServicesTokenRequestStart);
+        return SuperwallEvent._(type: EventType.adServicesTokenRequestStart);
       case 'adServicesTokenRequestFail':
-        return SuperwallPlacement._(
-            type: PlacementType.adServicesTokenRequestFail,
-            error: json['error']);
+        return SuperwallEvent._(
+            type: EventType.adServicesTokenRequestFail, error: json['error']);
       case 'adServicesTokenRequestComplete':
-        return SuperwallPlacement._(
-            type: PlacementType.adServicesTokenRequestComplete,
+        return SuperwallEvent._(
+            type: EventType.adServicesTokenRequestComplete,
             token: json['token']);
       case 'shimmerViewStart':
-        return SuperwallPlacement._(type: PlacementType.shimmerViewStart);
+        return SuperwallEvent._(type: EventType.shimmerViewStart);
       case 'shimmerViewComplete':
-        return SuperwallPlacement._(type: PlacementType.shimmerViewComplete);
+        return SuperwallEvent._(type: EventType.shimmerViewComplete);
       default:
         throw ArgumentError('Invalid placement type: ${json}');
     }